Vu Do Branch & User

+++ PawnShopBE/PawnShopBE.Core/Models/Branch.cs
@@ -15,7 +15,7 @@ namespace PawnShopBE.Core.Models
         public string BranchName { get; set; }
         public string Address { get; set; }
         public string PhoneNumber { get; set; }
-        public int Fund { get; set; }
+        public decimal Fund { get; set; }
         public DateTime CreateDate { get; set; }
         public DateTime? UpdateDate { get; set; }
         public int Status { get; set; }
